Refactor modules and components
