+++ wayne/stock_api_server/frontend/src/app/components/components.module.js
@@ -1,0 +1,13 @@
+import angular from 'angular';
+
+import navbar from './navbar/navbar.module';
+import hero from './hero/hero.module';
+import turnovers from './turnovers/turnovers.module';
+
+export default angular
+  .module('components', [
+    navbar,
+    hero,
+    turnovers
+  ])
+  .name;
