+++ wayne/stock_api_server/frontend/src/app/components/hero/hero.component.js
@@ -1,9 +1,7 @@
-const hero = {
+export default {
   template: require('./hero.html'),
   bindings: {
     // one-way bindings
     link: '<'
   }
 };
-
-export default hero;
