+++ wayne/stock_api_server/frontend/src/app/components/hero/hero.module.js
@@ -1,0 +1,7 @@
+import angular from 'angular';
+import hero from './hero.component';
+
+export default angular
+  .module('components.hero', [])
+  .component('hero', hero)
+  .name;
