--- wayne/stock_api_server/frontend/src/app/components/hero/index.js
@@ -1,3 +1,0 @@
-import hero from './hero';
-
-export default hero;
