--- wayne/stock_api_server/frontend/src/app/components/navbar/index.js
@@ -1,3 +1,0 @@
-import navbar from './navbar';
-
-export default navbar;
