+++ wayne/stock_api_server/frontend/src/app/components/navbar/navbar.component.js
@@ -1,11 +1,9 @@
 import './nabvar.css';
 
-const navbar = {
+export default {
   template: require('./navbar.html'),
   bindings: {
     // one-way bindings
     link: '<'
   }
 };
-
-export default navbar;
