+++ wayne/stock_api_server/frontend/src/app/components/navbar/navbar.module.js
@@ -1,0 +1,7 @@
+import angular from 'angular';
+import navbar from './navbar.component';
+
+export default angular
+  .module('components.navbar', [])
+  .component('navbar', navbar)
+  .name;
