+++ wayne/stock_api_server/frontend/src/app/components/turnovers/turnovers.module.js
@@ -1,0 +1,10 @@
+import angular from 'angular';
+
+import turnoversForm from './turnoversForm/turnoversForm.component';
+import turnoversList from './turnoversList/turnoversList.component';
+
+export default angular
+  .module('components.turnovers', [])
+  .component('turnoversForm', turnoversForm)
+  .component('turnoversList', turnoversList)
+  .name;
