+++ wayne/stock_api_server/frontend/src/app/components/turnovers/turnoversForm/turnoversForm.component.js
@@ -1,0 +1,3 @@
+export default {
+  template: require('./turnoversForm.html')
+};
