+++ wayne/stock_api_server/frontend/src/app/components/turnovers/turnoversList/turnoversList.component.js
@@ -1,9 +1,7 @@
-const turnoversList = {
+export default {
   template: require('./turnoversList.html'),
   bindings: {
     // one-way bindings
     turnovers: '<'
   }
 };
-
-export default turnoversList;
