+++ wayne/stock_api_server/frontend/src/app/components/turnovers/turnoversList/turnoversList.html
