--- wayne/stock_api_server/frontend/src/app/components/turnoversForm/index.js
@@ -1,3 +1,0 @@
-import turnoversForm from './turnoversForm';
-
-export default turnoversForm;
