--- wayne/stock_api_server/frontend/src/app/components/turnoversForm/turnoversForm.js
@@ -1,5 +1,0 @@
-const turnoversForm = {
-  template: require('./turnoversForm.html')
-};
-
-export default turnoversForm;
