--- wayne/stock_api_server/frontend/src/app/components/turnoversList/index.js
@@ -1,3 +1,0 @@
-import turnoversList from './turnoversList';
-
-export default turnoversList;
